Reorganize classes into packages
