+++ build.gradle
@@ -29,7 +29,7 @@ test {
 }
 
 application {
-    mainClassName = "seedu.duke.Gordon"
+    mainClassName = "Gordon"
 }
 
 shadowJar {
