+++ src/main/java/gordon/util/Difficulty.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package gordon.util;
 
 public enum Difficulty {
     Easy, Medium, Hard
