--- src/main/java/seedu/duke/Gordon.java
@@ -1,12 +1,0 @@
-package seedu.duke;
-
-import java.util.ArrayList;
-import java.util.Scanner;
-
-public class Gordon {
-    public static void main(String[] args) {
-        Cookbook mainCookbook = new Cookbook();
-        Parser parser = new Parser();
-        parser.parseMaster(mainCookbook);
-    }
-}
